requirements.txt
fabric>=1.8
bitbucket-api>=0.5
python-digitalocean>=0.7
